Final Commit
Edit functionality is 95% there. Something is wrong with the active
record portion and it is not upadating the DB.
Logout button does not work and I have no idea why.
Table displays hide when a modal or link is clicked. But reapear easily
when you click on one of the pageination links.

It was a good go :)