Fixed a bug where edit orginization wasnt working properly.
Found a new bug: when you edit User data, if you change the
organization they are appart of it does not update the
'organizations_id' in session, thus making it apear as if they arent
part of that org. I need to find a way to reset the session data.